+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to
 - The canonical format (`-Tcanon`) produces output that is stable across
   multiple invocations. #2193
 
+- `gvpack` now preserves input graph labels as labels on the packed component
+  subgraphs. #2415
+
 ## [16.1.0] – 2026-09-03
 
 ### Added